==> analysis/33_20230504_DM_DNAFISH_24hr/03_simplified_analysis-multiple.py <==
import skimage.io as skio
import pandas as pd
from skimage.morphology import disk, dilation, medial_axis
from skimage.measure import label, regionprops
import shared.image as ima
import numpy as np
import shared.dataframe as dat
import shared.math as mat
import math
import os

# INPUT PARAMETERS
# file info
master_folder = "/Users/xwyan/Dropbox/LAB/ChangLab/Projects/Data/20230504_analysis_DM_24hr/"
data_dir = "%sdata/" % master_folder
data_dir1 = "%sfigures/" % master_folder
output_dir = "%sfigures/" % master_folder

exp = '20230317_DM_24hr_DNAFISH'

row = 'G'
column = [3, 4, 5, 6, 7, 8, 9, 10, 11]
sample_lst = ['%s%s' % (row, x) for x in column]
# sample_lst = ['B4', 'B9', 'C7']

for sample in sample_lst:
    file_name = '%s_%s_%s_RAW' % (exp, sample[0], sample)
    img_hoechst_stack = skio.imread("%s/%s_ch01.tif" % (data_dir, file_name), plugin="tifffile")
    img_DNAFISH_stack = skio.imread("%s/%s_ch00.tif" % (data_dir, file_name), plugin="tifffile")

    n_nuclear_convex_dilation = 4
    local_size = 200
    rmax = 100


    def img_to_pixel_int(mask: np.array, img: np.array):
        index = [i for i, e in enumerate(mask.flatten()) if e != 0]
        out = list(map(img.flatten().__getitem__, index))
        return out


    data = pd.DataFrame(columns=['well', 'nuclear', 'FOV', 'n_nuclear_convex_dilation',
                                 'centroid_nuclear', 'area_nuclear', 'circ_nuclear', 'mean_int_nuclear',
                                 'mean_int_DNAFISH', 'n_ecDNA',
                                 'mean_int_ecDNA', 'total_area_ecDNA', 'total_area_ratio_ecDNA',
                                 'dis_to_hub_area',
                                 'nuclear_int', 'DNAFISH_int',
                                 'DNAFISH_seg_label', 'int_r_to_edge', 'int_r_to_center', 'int_relative_r'])

    for fov in range(img_hoechst_stack.shape[0]):
        print("Analyzing %s, fov %s" % (sample, fov))
        img_nuclear_bgc = img_hoechst_stack[fov, :, :]
        img_DNAFISH_bgc = img_DNAFISH_stack[fov, :, :]
        img_seg = skio.imread("%s/seg_tif/%s/%s_%s_seg.tif" % (data_dir1, sample, sample, fov), plugin="tifffile")
        img_ecSeg = skio.imread("%s/seg_tif/%s/%s_%s_ecseg1.tif" % (data_dir1, sample, sample, fov), plugin="tifffile")

        if n_nuclear_convex_dilation > 0:
            img_nuclear_seg = dilation(img_seg, disk(n_nuclear_convex_dilation))

        # measure
        # get local images
        nuclear_props = regionprops(img_nuclear_seg)

        for i in range(len(nuclear_props)):

            print("Analyzing %s, fov %s, nuclear %s/%s" % (sample, fov, i + 1, len(nuclear_props)))
            original_centroid_nuclear = nuclear_props[i].centroid
            position = ima.img_local_position(img_nuclear_seg, original_centroid_nuclear, local_size)
            local_nuclear_seg = ima.img_local_seg(img_nuclear_seg, position, nuclear_props[i].label)
            local_nuclear = img_nuclear_bgc.copy()
            local_nuclear = local_nuclear[position[0]:position[1], position[2]:position[3]]
            local_DNAFISH = img_DNAFISH_bgc.copy()
            local_DNAFISH = local_DNAFISH[position[0]:position[1], position[2]:position[3]]
            local_DNAFISH_seg = img_ecSeg.copy()
            local_DNAFISH_seg = local_DNAFISH_seg[position[0]:position[1], position[2]:position[3]]
            local_DNAFISH_seg[local_nuclear_seg == 0] = 0

            # basic measurements
            local_nuclear_props = regionprops(label(local_nuclear_seg), local_nuclear)
            local_DNAFISH_props = regionprops(label(local_nuclear_seg), local_DNAFISH)
            ecDNA_props = regionprops(label(local_DNAFISH_seg), local_DNAFISH)

            area_nuclear = local_nuclear_props[0].area
            perimeter_nuclear = local_nuclear_props[0].perimeter
            mean_int_nuclear = local_nuclear_props[0].intensity_mean
            mean_int_DNAFISH = local_DNAFISH_props[0].intensity_mean
            circ_nuclear = (4 * math.pi * area_nuclear) / (perimeter_nuclear ** 2)

            # ecDNA measurements
            n_ecDNA = len(ecDNA_props)
            centroid_ind_ecDNA = [ecDNA_props[i].centroid for i in range(len(ecDNA_props))]
            area_ind_ecDNA = [ecDNA_props[i].area for i in range(len(ecDNA_props))]
            area_ratio_ind_ecDNA = list(np.array(area_ind_ecDNA) / area_nuclear)
            total_area_ecDNA = sum(area_ind_ecDNA)
            total_area_ratio_ecDNA = sum(area_ratio_ind_ecDNA)

            mean_int_ind_ecDNA = [ecDNA_props[i].intensity_mean for i in range(len(ecDNA_props))]
            total_int_ind_ecDNA = list(np.array(area_ind_ecDNA) * np.array(mean_int_ind_ecDNA))
            total_int_ecDNA = sum(total_int_ind_ecDNA)
            mean_int_ecDNA = total_int_ecDNA / total_area_ecDNA if total_area_ecDNA != 0 else 0

            # distance from hub v2
            dis_to_hub_area_v2 = 0

            if n_ecDNA == 0:
                dis_to_hub_area_v2 = 0
            elif n_ecDNA > 1:
                ind_ecDNA = pd.DataFrame({'area': area_ind_ecDNA, 'total_int': total_int_ind_ecDNA,
                                          'centroid': centroid_ind_ecDNA})

                ind_ecDNA_sort_area = ind_ecDNA.copy().sort_values(by='area', axis=0, ascending=False,
                                                                   inplace=False,
                                                                   ignore_index=True)
                dis = []
                for k in range(len(ind_ecDNA_sort_area)):
                    dis_temp = 0
                    target_ecDNA = ind_ecDNA_sort_area.iloc[[k]]
                    rest_ecDNA = ind_ecDNA_sort_area.copy().drop(k)
                    current_ecDNA = pd.concat([target_ecDNA, rest_ecDNA], axis=0).reset_index(drop=True)
                    current_ecDNA['dis'] = \
                        [((current_ecDNA['centroid'][i][0] - current_ecDNA['centroid'][0][0]) ** 2 +
                          (current_ecDNA['centroid'][i][1] - current_ecDNA['centroid'][0][1]) ** 2) ** 0.5
                         for i in range(len(current_ecDNA))]

                    for ind in range(n_ecDNA - 1):
                        dis_temp = dis_temp + current_ecDNA['area'][ind + 1] * current_ecDNA['dis'][
                            ind + 1] / total_area_ecDNA
                    dis.append(dis_temp)
                ind_ecDNA_sort_area['dis'] = dis

                for ind in range(n_ecDNA):
                    dis_to_hub_area_v2 = dis_to_hub_area_v2 + ind_ecDNA_sort_area['area'][ind] * \
                                         ind_ecDNA_sort_area['dis'][ind] / total_area_ecDNA

            # radial distribution
            local_nuclear_centroid = local_nuclear_props[0].centroid
            _, local_edge_distance_map = medial_axis(local_nuclear_seg, return_distance=True)
            local_centroid_distance_map = ima.distance_map_from_point(local_nuclear_seg, local_nuclear_centroid)
            local_centroid_distance_map[local_nuclear_seg == 0] = 0
            local_edge_distance_map[local_nuclear_seg == 0] = -1
            local_relative_r_map = local_centroid_distance_map / (
                    local_centroid_distance_map + local_edge_distance_map)

            nuclear_int = img_to_pixel_int(local_nuclear_seg, local_nuclear)
            DNAFISH_int = img_to_pixel_int(local_nuclear_seg, local_DNAFISH)
            DNAFISH_seg_label = img_to_pixel_int(local_nuclear_seg, local_DNAFISH_seg)
            int_r_to_edge = img_to_pixel_int(local_nuclear_seg, local_edge_distance_map)
            int_r_to_center = img_to_pixel_int(local_nuclear_seg, local_centroid_distance_map)
            int_relative_r = img_to_pixel_int(local_nuclear_seg, local_relative_r_map)

            data.loc[len(data.index)] = [sample, i, fov, n_nuclear_convex_dilation,
                                         original_centroid_nuclear, area_nuclear, circ_nuclear, mean_int_nuclear,
                                         mean_int_DNAFISH, n_ecDNA,
                                         mean_int_ecDNA, total_area_ecDNA, total_area_ratio_ecDNA,
                                         dis_to_hub_area_v2,
                                         nuclear_int, DNAFISH_int, DNAFISH_seg_label, int_r_to_edge, int_r_to_center, int_relative_r]

    if not os.path.exists("%s/txt/" % output_dir):
        os.makedirs("%s/txt/" % output_dir)
    data.to_csv('%s/txt/%s_n%s.txt' % (output_dir, sample, n_nuclear_convex_dilation), index=False, sep='\t')

print("DONE!")
